[CHORE] coupon 도메인 패키지 domain 내부로 이전
