--- src/main/java/com/programmers/smrtstore/coupon/domain/entity/CouponPublicationType.java
@@ -1,5 +1,0 @@
-package com.programmers.smrtstore.coupon.domain.entity;
-
-public enum CouponPublicationType {
-    ALLOCATE,DOWNLOAD
-}
