--- src/main/java/com/programmers/smrtstore/coupon/domain/entity/CouponType.java
@@ -1,5 +1,0 @@
-package com.programmers.smrtstore.coupon.domain.entity;
-
-public enum CouponType {
-    CART,SHIPPING_PRICE
-}
