--- src/main/java/com/programmers/smrtstore/coupon/presentation/req/SaveCouponRequest.java
@@ -1,4 +1,0 @@
-package com.programmers.smrtstore.coupon.presentation.req;
-
-public class SaveCouponRequest {
-}
