--- src/main/java/com/programmers/smrtstore/coupon/presentation/res/OrderCouponResponse.java
@@ -1,4 +1,0 @@
-package com.programmers.smrtstore.coupon.presentation.res;
-
-public class OrderCouponResponse {
-}
