+++ src/main/java/com/programmers/smrtstore/domain/coupon/domain/entity/CouponType.java
@@ -1,0 +1,5 @@
+package com.programmers.smrtstore.domain.coupon.domain.entity;
+
+public enum CouponType {
+    CART,SHIPPING_PRICE
+}
