+++ src/main/java/com/programmers/smrtstore/domain/coupon/domain/entity/CustomerManageBenefitType.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.coupon.domain.entity;
+package com.programmers.smrtstore.domain.coupon.domain.entity;
 
 public enum CustomerManageBenefitType {
     ALL,NEW,TALK_FRIEND,REPURCHASE
