+++ src/main/java/com/programmers/smrtstore/domain/coupon/domain/entity/ProductCoupon.java
@@ -1,4 +1,4 @@
-package com.programmers.smrtstore.coupon.domain.entity;
+package com.programmers.smrtstore.domain.coupon.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
