+++ src/main/java/com/programmers/smrtstore/domain/coupon/infrastructure/ProductCouponJpaRepository.java
@@ -1,6 +1,6 @@
-package com.programmers.smrtstore.coupon.infrastructure;
+package com.programmers.smrtstore.domain.coupon.infrastructure;
 
-import com.programmers.smrtstore.coupon.domain.entity.ProductCoupon;
+import com.programmers.smrtstore.domain.coupon.domain.entity.ProductCoupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductCouponJpaRepository extends JpaRepository<ProductCoupon,Long> {
