+++ src/main/java/com/programmers/smrtstore/domain/coupon/infrastructure/UserCouponJpaRepository.java
@@ -1,6 +1,6 @@
-package com.programmers.smrtstore.coupon.infrastructure;
+package com.programmers.smrtstore.domain.coupon.infrastructure;
 
-import com.programmers.smrtstore.coupon.domain.entity.UserCoupon;
+import com.programmers.smrtstore.domain.coupon.domain.entity.UserCoupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserCouponJpaRepository extends JpaRepository<UserCoupon,Long> {
