+++ src/main/java/com/programmers/smrtstore/domain/coupon/presentation/req/CreateCouponRequest.java
@@ -1,0 +1,4 @@
+package com.programmers.smrtstore.domain.coupon.presentation.req;
+
+public class CreateCouponRequest {
+}
