+++ src/main/java/com/programmers/smrtstore/domain/coupon/presentation/req/OrderCouponRequest.java
@@ -1,0 +1,4 @@
+package com.programmers.smrtstore.domain.coupon.presentation.req;
+
+public class OrderCouponRequest {
+}
