+++ src/main/java/com/programmers/smrtstore/domain/coupon/presentation/req/SaveCouponRequest.java
@@ -1,0 +1,4 @@
+package com.programmers.smrtstore.domain.coupon.presentation.req;
+
+public class SaveCouponRequest {
+}
