+++ src/main/java/com/programmers/smrtstore/domain/coupon/presentation/res/OrderCouponResponse.java
@@ -1,0 +1,4 @@
+package com.programmers.smrtstore.domain.coupon.presentation.res;
+
+public class OrderCouponResponse {
+}
