+++ src/main/java/com/programmers/smrtstore/domain/coupon/presentation/res/SaveCouponResponse.java
@@ -1,0 +1,4 @@
+package com.programmers.smrtstore.domain.coupon.presentation.res;
+
+public class SaveCouponResponse {
+}
